refactor: organize MySQL CLI tests
